requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# TMDB API
httpx==0.26.0

# File parsing
guessit==3.8.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Utilities
watchdog==3.0.0
apscheduler==3.10.4

# Frontend serving
aiofiles==23.2.1
jinja2==3.1.3
